Fix other-requirements.txt
With the switch to bindep based packages the other-requirements.txt file
is used. It contains an error: The libxslt devel is called
libxslt1-dev on Ubuntu Trusty.

Also, add mysql-server and mysql-client packages, as well as postgresql
packages, those are needed by gate-requirements-python27.

Add pypy packages for pypy gate.

This is a backport of a9b887b4020431e177fd5f62092ef636a4ec398.

Change-Id: Ie143ad862b88034ef47470b85fd02c1d040415a7