Update patch set 2
Patch Set 2:

sorantis asked in #openstack-release whether we could make an exception for this patch to add kingbird to the requirements list. That's not what this patch does, though. It adds kingbird to the list of projects that receive automatic requirements syncs. There won't be any significant requirements changes, so it should be possible to just sync them once by hand if you need them synced. We don't usually add services to the global requirements list, so unless there is a client library for kingbird it shouldn't be necessary to modify any other files here and it will be safe to wait to apply this change until we unfreeze the requirements repository.

Patch-set: 2