Bump python-troveclient to 2.1.0
Trove dropped support for the slave_of argument when creating
instances in 0c94a5bae0aacff7fde9a9742b7d526096e51767.

python-troveclient dropped slave_of support in
a5234f0a07230e0d8195687e033adfe4cc0ad522 which was released
in 2.1.0 (since it's a backward incompatible change, it should
have been a major version bump to 3.0.0, but that's a different
issue).

Since the trove-api no longer supports slave_of, versions of
the client before 2.1.0 won't work with the mitaka version of
the API, at least if they try sending slave_of in the request.

So this signals that the minimum required version of the client
that should be used in mitaka is 2.1.0 so people using the
client at least can't use slave_of in the python API in the
client.

Note that an attempt was made to revert the client change
in I9b3ec6c101a89f30e61fc683808506c404e5474f but because
of the backward incompatible API change in the server, the
revert fails. So unless the API change is reverted, which it
probably won't be, we need to do the next best thing possible
which is move everything forward.

Change-Id: I7ec4b07ec2f6b79e26a846b801514c1568646b87
Related-Bug: #1538506