Update patch set 2
Patch Set 2:

Given the Jenkins failures I think you probably have to rebase this?

I share Doug's concerns on this, please provide the explanation in the commit message. I'm not privy to the gentoo build process for openstack and I'm not sure if we should be making this change just based on that (would be good to get an ack from zigo for example).

Patch-set: 2
Label: Verified=0