Update patch set 3
Patch Set 3: Code-Review-1

upper-constraints is what community uses to test, and post release, the idea is not to bump it unless there is some critical reason to do so.

The requirements range exists for a reason. But moving our testing to pins that are distro specific becomes a new giant negotiation that we were actually trying to avoid.

Patch-set: 3
Reviewer: Gerrit User 2750 <2750@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1