Update patch set 1
Patch Set 1:

(1 comment)

I'm hesitant to bump the minimum required versions of dependencies in a stable branch unless we absolutely need to, e.g. you can't start a service without it.

I think the upper-constraint changing to 0.4.1 is OK, but I'm not really OK with changing the minimum at this point in stable.

Patch-set: 1
Label: Verified=0