Update patch set 3
Patch Set 3:

Is the library actively maintained?
------>yes,it is stable now. Ref: https://pypi.python.org/pypi/crc16/0.1.1
Is the library good code?
------->yes,we test it in our dragonflow project,it is a good code
Is the library python 3 compatible?
-------->yes,ref https://pypi.python.org/pypi/crc16/0.1.1
Is the library license compatible?
-------->yes, License is  LGPL 
ref https://pypi.python.org/pypi/crc16/0.1.1

Is the library already packaged in the distros we target (Ubuntu latest / Fedora latest)?------------>I am not sure about this. how to check it ?
Is the function of this library already covered by other libraries in global-requirements.txt?-------> yes, I check myself no lib covers crc16 lib

Patch-set: 3
Label: Verified=0