Update patch set 1
Patch Set 1: Code-Review+2

The stable process has been followed and while this isn't a critical/security bug it does mean that the tools work on non POSIX systems.

I'm a little nervous as we don't have any tests for this code, but parts of it are executed in dsvm jobs.  Sean is a requirements-core so lets do this.

Holding off +W until the stable/mitaka change has landed

Patch-set: 1
Label: Code-Review=+2