Update patch set 1
Patch Set 1:

OK, so blacklisting the bad version is fine, even if we aren't using it in upper-constraints. It would at least be in global-requirements in case we ever did try to bump upper-constraints in stable/mitaka.

I'm -1 on this change though because of the commit message, you should be using git cherry-pick -x <hash> and leave that in the commit message, plus the change ID should be the same as the original change you're cherry picking from, basically what Tony has here:

https://review.openstack.org/#/c/308733/

So I'm going to +2 that one.

Patch-set: 1