Add spec-cleaner to list of global-requirements
spec-cleaner is a program to enforce a specfic style for
RPM .spec files. It's similar to pep8 for python code, but
spec-cleaner doesn't have a specification. The library is
only used for testing and is not needed while running any
OpenStack component.

The rpm-packaging project.

Yes. It's available on https://github.com/openSUSE/spec-cleaner and
a couple of different people are working on it. The module as also
a big unittest suite.

Available and also tested with unittests.

Yes. It's BSD licensed.

Change-Id: I41ebffcb6646c3c44e469f6393a1aef86bcd927e