Update patch set 1
Patch Set 1:

I want that this change is merged.
But in another change[1], I'm requested to propose it for master first to patch for stable branch by member of stable branch liaison.
I think that it is unnecessary to maintain the master branch when we plan to remove a project.
So I will stop this change until it clears.

[1] https://review.openstack.org/#/c/308084/
    https://review.openstack.org/#/c/308107/

Patch-set: 1