Update patch set 1
Patch Set 1:

> That makes sense, if this changes to allow 18.4 it will probably
 > need to go to master (or not) and then mitaka (or not) and liberty.

Jesse's change as proposed does NOT allow 0.18.4.  The alternate approach which I outlined to clarify what Jesse is proposing:

 eventlet>=0.16.1,!=0.17.0,!=0.18.0,!=0.18.1,!=0.18.2,!=0.18.3

Does allow 0.18.4 and causes the upgrade problem which is why I prefer to take this review as is or do nothing.  It's not like we're going to release the world in the next month

Patch-set: 1