Change urllib3 version to urllib3>=1.15
Urllib3 1.15 introduced support for chunked transfer encoding, which
Tempest used to do with some custom code using the low-level httplib
module. Instead, it would be nice to rely on Urllib3 to do this.

Urllib3 release note:
https://github.com/shazow/urllib3/blob/master/CHANGES.rst#115-2016-04-06

requests is also bumped as part of this commit because it bundles
urllib3 and since most distros unbundle it the versions need to be
kept in sync.

Change-Id: I255e70e6b680e3b61264b5cf49621c40f460ca8e