Update patch set 1
Patch Set 1: Code-Review-1

Negative for the following reason from the readme

If the change was not proposed by the OpenStack CI bot, and does not include a global-requirements.txt change, then it should be rejected: the CI bot will generate an appropriate change itself. Ask in #openstack-infra if the bot needs to be run more quickly.

Otherwise the change may be the result of recalculating the constraints which changed when a global-requirements.txt change is proposed. In this case, ignore the changes to upper-constraints.txt and review the global-requirements.txt component of the change.

Patch-set: 1