Update patch set 1
Patch Set 1:

As a packager this is fine for Gentoo at least, since we support multiple versions of packages in our tree.  I'd like to see what other distros thing, also the submitter could go though the distros and document them in the PR (the links are in the readme).

`[I] dev-python/requests
     Available versions:  2.7.0^t ~2.8.0^t 2.8.1^t{tbz2} ~2.9.1-r2^t {test PYTHON_TARGETS="pypy pypy3 python2_7 python3_3 python3_4 python3_5"}`

Patch-set: 1