Update patch set 1
Patch Set 1:

Yeah, that's kinda unreasonable especially considering that newton development started ~3 weeks ago. The newer urllib3 is going to be a requirement for tempest to work correctly. (it's also needed to fix some bugs we're working through right now too)

This is also all because you choose to unbundle a dependency that is bundled upstream. You wouldn't have an issue if you just left requests alone in rhel. As a compromise I think we can bump urllib3 and requests together, but holding off here for 6 months isn't gonna happen.

Patch-set: 1