Update patch set 1
Patch Set 1:

> This is also all because you choose to unbundle a dependency that is bundled upstream.

Or upstream chose to bundle urllib3 despite common sense. You're not the one who will have fun fixing hundred packages because of a single CVE in urllib3.
Complain to requests maintainers for asking us unreasonable requests like enforcing that we ship the same version of urllib3 than the one used in requests.

> the newer urllib3 is going to be a requirement for tempest to work correctly. (it's also needed to fix some bugs we're working through right now too)

"which Tempest used to do with some custom code using the low-level httplib module. Instead, it would be nice to rely on Urllib3 to do this." => from my understanding, there is already code doing it in tempest, so it does not seem like a hard requirement.

> As a compromise I think we can bump urllib3 and requests together, but holding off here for 6 months isn't gonna happen.

This is not compromise, you're just forcing your whim on us. 
And yes, urllib3 and requests lower bounds must be bumped together period.

> I mean RHEL7/Centos 7 is not going away in 6 months

I'm not saying we won't upgrade it but it's not easy. Again requests is a very popular package, so updating it may break other critical packages elsewhere (e.g ceph, some yum plugins)
And we don't wait Newton release to start working on packages, actually we build packages for every commit pushed.

Patch-set: 1