Update patch set 1
Patch Set 1:

Alan, when we created stable/liberty the bot that proposed updates to u-c was still running on stable and two of those got approved. We disabled that awhile ago for stable branches, and it's not a problem on stable/mitaka. Rather than revert the changes outright, which weren't clean, we've just been rewinding these as needed, like in this case.

Patch-set: 1