Update patch set 2
Patch Set 2: Code-Review+1

This is a welcome change, tripleo-common is a requirement for python-tripleoclient which is tracked by g-r.
Note that 2.0.0 is the latest one, but since it's part of tripleo and we're speaking about master, I'm fine with this as a lower bound.

Patch-set: 2
Reviewer: Gerrit User 10384 <10384@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1