Update patch set 2
Patch Set 2:

So.. while I respect that this is obviously the project maintainers choice, I want to make sure you are aware of all the potential consequences before you jump... In particular, we traditionally linked the "tc-approved-release" concept to a proven-coinstallable set of projects (using the requirements repository to enforce it). So this might result in Ceilometer losing the tc-approved-release tag if the TC ends up confirming this old 'approved-release=can be integrated=must use requirements' rule. The tag is only used for candidates for defcore programs those days, so I guess it doesn't matter much, but I thought I'd rather point out a potential consequence of this change before I W+1 it.

Patch-set: 2