Update patch set 1
Patch Set 1:

The known break is around paramiko.RSAKey

No other project in codesearch has the signature that is a problem - http://codesearch.openstack.org/?q=paramiko.RSAKey&i=nope&files=&repos=

All the other projects only use functions that are shown to work with the new version (because cinder / tempest use them, and are thus tested).

I think that as long as the sahara and ironic tests come in we have a pretty high confidence level that this is ok. And it can be reverted should it become an issue.

Patch-set: 1