Update patch set 1
Patch Set 1:

Why don't we set this up as depending on the patches for the sahara and ironic, to avoid breaking them until those have landed. Then we can merge this and update the requirement, and as long as that doesn't uncover anything nova can remove its version checking code.

Patch-set: 1