Update patch set 1
Patch Set 1:

the context for this is the following change [1] which failed with a script error in project-requirements-change.py

It's looking for the "extras" field in the Requirement object, which isn't there on this branch.  Although I'm not sure extras is required on liberty, I think the object needs to define it if the requirements checking script is to work everywhere?

[1] http://logs.openstack.org/43/308143/2/check/gate-networking-fujitsu-requirements/ea229d9/console.html

Patch-set: 1
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0