Update patch set 2
Patch Set 2:

@Doug - thanks for the check!

Given latest tosca-parser (0.5.0) is already in pypi, Tacker gate is broken even without this change. We are sort of in a rock and a hard place. Sahdev / tosca-parser is promising to maintain backward compatibility going forward. For now we need this and a follow on heat-translator 0.5.0 release to have Tacker gate unblocked.

Patch-set: 2