Update patch set 1
Patch Set 1:

> I'd rather see Iff5fb3180855de7adb3399f6be16bedc8543b4ec
 > backported, but this will fix the issue too, and prevent it from
 > happening in many other spots

Working on it :)

I think I'm going to publish both approaches and see which is better/more acceptable.

I've made 1581287 a duplicate of 1568527 as it really is the same issue.

A look at master, stable/mitaka and stable/liberty for users of tempest_lib makes it look like it's only novaclient that's using this.  I'm mildly surprised to see that tempest_lib doesn't require keystoneclient but that's an entirely different conversation.

Patch-set: 1