Update patch set 1
Patch Set 1: Code-Review-1

Roland,

Thanks for looking into this.

I see two references:
http://codesearch.openstack.org/?q=kafka-python&i=nope&files=&repos=monasca-api,monasca-common

Both those projects do *not* subscribe to check-requirements job:
http://git.openstack.org/cgit/openstack-infra/project-config/tree/zuul/layout.yaml#n7386
http://git.openstack.org/cgit/openstack-infra/project-config/tree/zuul/layout.yaml#n7414

And are not listed as subscribed to g-r/u-c process:
http://git.openstack.org/cgit/openstack/requirements/tree/projects.txt

We can revisit this when you get the rest of the Monasca libraries in and the only one left is kafka. Hopefully by that time, you can update to latest kafka as well. It's a bit unfair to hold other people up like oslo.messaging that want to move faster and fix performance issues with newer library.

Thanks,
Dims

Patch-set: 1