Update patch set 1
Patch Set 1:

Having the requirement set to > 0.9.5 doesn't make any sense as 0.9.5 is not compatible with 1.0.0.

There are a number of problems with the 1.X releases of the python-kafka library. We shouldn't be upgrading to anything in 1.X at this time and we've been working with the developers of the python-kafka library to resolve them. We shouldn't be introducing unstable Python libraries into the OpenStack global requirements/

Monasca is an OpenStack project. We are trying to get all of our stuff merged as well as the global requirements. It is a work in progress and we are getting close to completing that. I understand the technical arguments, you aren't in global requirements, so therefore you don't count yet. I think in this case we need to consider not just the technical argument, but what is the right answer.

How are we holding up any project? There are performance problems in the python-kafka library. They should be resolved prior to updating and adding them to OpenStack global requirements. You are holding up the Monasca project and rest of the Monasca eco-system right now.

See the review from https://review.openstack.org/#/c/282147, the review set the upper limit on requirements as well as my abandoned review at, https://review.openstack.org/#/c/282053/.

Patch-set: 1
Reviewer: Gerrit User 2419 <2419@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0