Update patch set 1
Patch Set 1:

Davanum, I understand the oslo team is trying to make progress. I really don't want to be holding projects back. 

If all of Monasca was already in the global requirements, then the change to remove the upper constraint would not have been allowed because it would have been caught by gate check. That would have prevented this change from being merged.

However, Monasca isn't in the global requirements. We've been trying to get there and are very close. So, should we allow the latest python-kafka libraries to the significant detriment of one team to allow another team to do future development. That doesn't seem like the right answer.

This change breaks Monasca. Our CI jobs are not able to run. It isn't a trivial matter to upgrade to the latest python-kafka library. We have already been looking into that and have run into a number of bugs and significant performance regressions.

Patch-set: 1