Update patch set 1
Patch Set 1: Code-Review+2 Workflow+1

Roland,

Looks like Monasca team needs a quick primer on the requirements process.

Step #1: Submit a patch to this file at the following locations to add a "check-requirements" job
http://git.openstack.org/cgit/openstack-infra/project-config/tree/zuul/layout.yaml#n7386
http://git.openstack.org/cgit/openstack-infra/project-config/tree/zuul/layout.yaml#n7414

This step adds a CI job to your review process, when the requirements.txt in your project differs from the one in global-requirements.txt it will complain loudly and fail your review. This ensures that your project requirements.txt will not have anything other than global-requirements.txt

Step #2: Submit a patch to this file
http://git.openstack.org/cgit/openstack/requirements/tree/projects.txt

This step ensures that when the nightly proposal bot/script runs, it will create a fresh review in your project queue with any changes that were made to global-requirements.txt. 

We won't accept a patch for projects.txt in step #2 without seeing a review for step #2.

Once you have Step #1 and Step #2, your project is essentially subscribed to the global requirements process. If we break you, then we revert stuff or work with you to move forward. If you see the list of projects in projects.txt, essentially we are guaranteeeing that we won't break any of them.

Now, The only project in the Monasca eco system that is in projects.txt and global-requirements.txt is python-monascaclient

Step #3, way ahead in the future is adding other monasca projects to global-requirements.txt. This is what you are talking about but maybe missing the fact that you can do Step #1 and Step #2 *RIGHT NOW*. If you had done Step #1 and #2, the requirements team is obligated to revert immediately, not otherwise.

As mentioned earlier, if it is onerous to do Step #1 and #2, then you should consider making sure that none of the jobs you run are subject to constraints. There are several project that don't like and don't want to be part of the requirements process.

So, in the hope that you can do Step #1 and #2 for all Monasca projects next week, i'll approve this revert late in the evening on a friday night :) I'll check with the oslo.messaging team working on Kafka for more details as well early next week. If Monasca team and oslo.messaging team can coordinate and get things going it will be the right thing to do. 

Thanks,
Dims

Patch-set: 1
Label: Code-Review=+2
Label: Workflow=+1