Update patch set 2
Patch Set 2:

Looking a bit further, it seems that only os-cloud-config is still in a project's requirements.txt (python-tripleoclient). The others are referenced in some projects, but the requirements are not explicitly stated.

This is what I am looking at to check.

http://codesearch.openstack.org/?q=os-apply-config&i=nope&files=requirements.txt&repos=

http://codesearch.openstack.org/?q=os-cloud-config&i=nope&files=requirements.txt&repos=

http://codesearch.openstack.org/?q=os-net-config&i=nope&files=requirements.txt&repos=

http://codesearch.openstack.org/?q=os-refresh-config&i=nope&files=requirements.txt&repos=

(side note, I didn't know about the codesearch, this is cool! Thanks Dirk)

Patch-set: 2