Update patch set 3
Patch Set 3:

I'd still like to understand the intent here a bit more, regardless of if we're missing some stuff in projects.txt, all of these projects *are* still used by tripleo.

Also, os-apply-config and os-refresh-config are requirements for heat software deployment features, so they are an indirect heat requirement (albeit optional if you decide not use use the deployment features).

If the drive is to remove things, that's fine, but I'm not clear on why we're removing things that are actually used and required in more than one project, or if that has implications if we then add them to requirements.txt (e.g where evidently they are missing) in the various places that require them.

Patch-set: 3