Restrict falcon>=1.0.0
Recently falcon just released a new version 1.0.0, which drop a
support of hooks for rest api which is used in Zaqar. Now Zaqar
has fix it to use the middleware mechanism of falcon, which can't
work with falcon version<1.0.0. Now in OpenStack community only
Zaqar is using falcon, so the change is safe.

Change-Id: I027f9c0c4047c5f0ff9e20bbce7cbbf9be1ccbb1