Update patch set 8
Patch Set 8:

The thing I'm worried about with lack of support is having a version that continues to work with newer versions of Python 3 (this one only claims 3.2) and any packaging tool changes that come along from setuptools. If you have tested with 3.4 yourself and it works, we can add the dependency, but be aware that projects using unmaintained dependencies will have to address issues that come up themselves.

Patch-set: 8