Update patch set 2
Patch Set 2:

OK, so it sounds like the release is likely to be API compatible.

The other question is whether we want to go raising the constraints like this at all in a stable branch. Because what you're really doing is raising the minimum required version, but you haven't edited the range in global-requirements.txt.

Patch-set: 2