Update patch set 3
Patch Set 3:

> > 2) Do the requirements removal proposed here and just state that
 > we manage dependencies in the package specs, which is what we've
 > been doing.
 > 
 > I suppose this is workable, but the implication then is e.g
 > tripleo-common is only consumable via delorean packages, which I
 > guess aligns with our current testing model, but it is released on
 > pypi too:
 > 
 > https://pypi.python.org/pypi/tripleo-common/2.0.0
 > 
 > And removing these from g-r would appear to imply that we cannot
 > fix the requirements provided there to match reality?

It's not that we can't fix it, it's that I'm not convinced we can keep them in sync long-term if we're not testing from pip installs.  I don't particularly like the idea of not having valid requirements for something released on pypi, but unless someone signs up to make sure it keeps working I don't see how it's going to happen.

Patch-set: 3