Update patch set 1
Patch Set 1: Code-Review+1

mock===2.0.0 is what we have in upper-constraints.txt so for the most part this should be fine.

We'll *still* need to keep an eye out for this in stable backports (and introduce a "conflict" with the master branch but stable is already a special snowflake so it shoudl be fine.

Patch-set: 1
Reviewer: Gerrit User 12898 <12898@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1