Update patch set 2
Patch Set 2: Code-Review-1

AFAICT the dependancy chain is:

 python-neutronclient>=2.6.0 in ./.tox/venv/lib/python2.7/site-packages (from ceilometer==5.0.3)
 cliff!=1.16.0,!=1.17.0,>=1.15.0 in ./.tox/venv/lib/python2.7/site-packages (from python-neutronclient>=2.6.0->ceilometer==5.0.3)

the liberty version of ceilometer doesn't need python-gnocchiclient

(Is there a bug we can look at to verify this?)

What we're seeing is a side-effect of not capping in global-requirements and relying on constraints to enforce the "cap" for us.

ceilometer 5.0.3 (liberty) has an uncapped requirement on python-neutronclient (4.2.0 [master]) which then pulls in cliff 2.0.0 and I imagine things get pretty messed up at that point.

I'm -1 on this until we understand this a little better, as it seems like we're essentially bumping the minimum for any service co-installed with ceilometer.

While in this case 2.0.0 is probably okay we need to be more sure.

Julien, I assume this is holding up the telemetry gate.  Any chance you can attend http://eavesdrop.openstack.org/#Requirements_Team_Meeting to work through this as quickly as possible?

Patch-set: 2
Reviewer: Gerrit User 12898 <12898@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1