Update patch set 2
Patch Set 2:

"You're saying that gnocchiclient requires this version, though. Which means that whatever is relying on gnocchiclient requires this version. Which means it's a new minimum."

So that means that the list of requirements managed by this repository is the list of the whole Python world we depend on it? That means that if e.g. gnocchiclient requires cliff 2.0.0 then the bot will propagate that dependency on all OpenStack project? That means if anything in the Python world has a de pendency, we propagate that to ALL projects of OpenStack? I knew this repository was a bad idea, but now it's starting to be completely crazy if what this it means.

Don't get me wrong, I'm ok to change the patch to make whatever it takes so we can fix Ceilometer gate, but it's really time to realise this repo idea is terrible. We're doing terrible work, losing time.


"the liberty version of ceilometer doesn't need python-gnocchiclient"


Ceilometer installs Gnocchi via devstack and Gnocchi needs Gnocchiclient, which requires cliff >1.16

Patch-set: 2