Update patch set 2
Patch Set 2:

I'll try to reply to everyone, but this is getting long :)

Sorry for the tone Tony, but we have incidents due to this repository on a monthly basis for year. It solved 0 issue so far, but create new ones every month, so we're really getting frustrated. Recently, it's especially the way devstack forces that upon  every project that strikes us indeed.

There's no bug open but the initial issue hit by ZQF is at https://review.openstack.org/#/c/315465/
I was not able to attend the meeting as you noticed. :)

Dims: I don't think there's way to have project unconstrained unfortunately

Patch-set: 2