Update patch set 1
Patch Set 1:

We want something like this to unblock the ceilometer liberty gate. I think we should add it to the global-requirements.txt as well, since it was technically a requirement but not declared as such. That would give us a minimum range, and the constraint entry would serve as the cap to prevent dragging cliff 2.0.0 in.

Patch-set: 1
Label: Verified=0