Bump up oslo messaging version
Commit 990d894eaf0d18941b56d89bff2bca51de375640 has fixed issue with
message acknowledgement where the messages are acknowledged only when
it is being processed and not when it is being queued. If the serving
process crashes, the messages can be recieved by other peer processes or
when the dead process restarts.

This is needed in Heat where we don't want the converge messages to be
lost when a heat-engine fails/crashes.

Change-Id: I059b766dea16aa299b58ada07ed3a398a9a164a1