Bump up oslo messaging version
Change I5f23e23644e90919cb67f81fc306ee85c5e09974 has fixed issue with
message acknowledgement where the messages are acknowledged only when
it is being processed and not when it is being queued. If the serving
process crashes, the messages can be recieved by other peer processes or
when the dead process restarts.

This is needed in Heat where we don't want the converge messages to be
lost when a heat-engine fails/crashes.

Change-Id: I059b766dea16aa299b58ada07ed3a398a9a164a1