Update patch set 1
Patch Set 1:

the oslo.messaging include a change[1] which constraint the amqp version to <2.0. in amqp-2.0 and amqp-2.0.1, there is an issue[2] that will cause amqp cannot be correctly installed in python3.4 environment. this issue has been broken the py34 jenkins job[3]. so this should be merged ASAP and then bump the version of oslo.messaging to >=5.2.0 in requirements.txt

any thoughts ?


[1] https://review.openstack.org/#/c/321704/
[2] https://github.com/celery/py-amqp/issues/91
[3] http://logs.openstack.org/41/323641/4/check/gate-ceilometer-python34/cf3f043/console.html

Patch-set: 1