Add pycodestyle, the new package name for pep8
pep8 has been renamed to pycodestyle. Adding this will allow projects to
transition over, and then pep8 can be removed.

  Is the library actively maintained?

Yes, as it is a continuation of the pep8 project it is being maintained by the
same developers.

https://github.com/PyCQA/pycodestyle/pulse/monthly

  Is the library good code?

Yes, it seems reasonable. It is as good, or better than the pep8 library that
we already use given it is the same code but with additional fixes and updates.

  Is the library python 3 compatible?

Yes. Like pep8 is continues to support both 2 and 3.

  Is the library license compatible?

Yes, it is MIT like pep8.

  Is the library already packaged in the distros we target (Ubuntu latest / Fedora latest)?

I don't believe it is yet, other than under the pep8 name. Given this is a code
quality tool and not something that will be deployed I don't think that is an
issue.

  Is the function of this library already covered by other libraries in global-requirements.txt?

Yes, pep8 has exactly the same functionality but it replaces pep8. Transition
should be as simple as replacing the usage of pep8 in the Tox.ini files for
pycodestyle. This is will be a very low impact and relatively easy transition.

Change-Id: I509c8a8713cc8f09a95e5df592c945e98d521555