Add tqdm progress metering library
This will be quite handy for usage in kolla so that the
image building can output a progress bar (instead of
being mostly silent)

Is the library actively maintained?

Yes, last updated on 2016-05-30

Is the library good code?

Seems well written and commented.

Is the library python 3 compatible?

Yes

Is the library license compatible?

Yes

Is the library already packaged in the distros we target?

Does not currently appear to be (but a google search
did show that there are deb packages and fedora packages
that may just not be in the normal repos).

Is the function of this library already covered by other
libraries in global-requirements.txt?

Not as far as I know.

Change-Id: Ib5e1542131524f66e2d5154956a23111902c2b19