Update patch set 12
Patch Set 12:

So Blazar does seem to subscribe the the requirements process:

 http://git.openstack.org/cgit/openstack-infra/project-config/tree/zuul/layout.yaml#n4742
 https://review.openstack.org/#/q/topic:openstack/requirements+project:openstack/blazar

I have a WIP patch to test the transition.  Clearly it needs to be cleaned up a lot before it'll merge  but it passes local testing so that's a bonus :)

While I'm touching blazar I'll add u-c support.

The WIP patch is at: I777424370dc02c3164803a63331a1befb8531de5

Patch-set: 12