Update patch set 2
Patch Set 2: Code-Review-1

In the past we've had security fixes in oslo libraries for but didn't bump the minimum required dependency in requirements.txt because we said that people could just pick up that version when reading the OSSA. I don't have a strong opinion either way - obviously it's easier to see that you need a higher minimum version when the requirements change.

There was a discussion about this last October:

http://lists.openstack.org/pipermail/openstack-dev/2015-October/076381.html

The consensus toward the end of the thread is you need to backport the fix to the stable branches of the library and release it, and make that known in the OSSA, but not bump the minimum required version since not everyone wants to pick up the newer version along with it's transitive dependencies and whatever other untested changes it brings along as part of that minimum.

So a soft -1 on this given the precedent.

Patch-set: 2