Update patch set 2
Patch Set 2:

After talking with Sean I realize now the nova change requires passing a new parameter to execute in processutils, so we have to have the minimum version bump for the dependency.

The outstanding question I have then is if we require 3.7.1 or 3.8.0. I prefer 3.7.1 to avoid pulling in other things from 3.8.0 that we didn't intend to be in mitaka (since 3.8.0 is a newton release).

Patch-set: 2