Update patch set 2
Patch Set 2:

Actually we don't technically need the min version bump, processutils.execute takes kwargs and prlimit is already there, it was added in version 3.4. The thing is if you didn't have the right version you wouldn't have the security fix, but that's what the OSSA is for, which I assume we assume people read and care about if they care about security fixes for their cloud.

Patch-set: 2