Update patch set 5
Patch Set 5:

> This isn't forcing projects to swtich [sic], just providing it for those that wish to use it.

Except there's no value in providing it to anyone to use. Everyone uses Flake8 for linting. There are numerous checks in pycodestyle 2.0 that exist in pep8 1.7.0 but did not exist before then which will cause every project that adopts it to have new errors because the checks are not yet at a stable point.

Until Flake8 updates to allow *either* pep8 *or* pycodestyle, people will have to add this to their gate job to check for linting problems separately and I doubt anyone really wants to extend those jobs by running all but a handful of checks twice (and the second time would be significantly slower than with Flake8 given the way Flake8 works).

So unless projects are *actively* trying to add pycodestyle to their gate, I still fail to see *any* value in adding this now.

Patch-set: 5