Update patch set 4
Patch Set 4: Code-Review+2

One of the main reasons we have this fleet of libraries, and that we have stable branches for them, is so that we can reduce the complexities in the projects. 

if oslo.concurrency > X in project code defeats so many things we are trying to do here. And pushes all kinds of new complexity back to the projects.

I feel very strongly this should be a bump. If distro folks want to object, they can. Or they can patch it out. But for upstream users of any of this code they really do need this version to get this security fix.

Patch-set: 4
Reviewer: Gerrit User 2750 <2750@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2