Raise oslo.policy requirement to 1.9.0
Oslo.policy 1.9.0 adds the ability to register default policy rules from
within a project. Once all rules are registered it is not necessary to
have a policy file specifying defaults, and a sample file can be
generated from the registered rules.

Nova would like to take advantage of this which requires bumping the min
version to 1.9.

Change-Id: Ibb53195b22b88cce13093b7b5b791c4a3dbb70e8