Add selector to requirements
selector is a simple WSGI routing and delegation library that is
going to be used in the new placement API currently being built
in nova. selector is more lightweight, declarative, and
inspectable than the Routes library in common use throughout
OpenStack. There's an explicit goal in the placement API to
"use different stuff". selector is just one part of that.

selector's one dependency, resolver, is also added to
upper-constraints.

For the new requirements questions:

Is the library actively maintained?

  There hasn't been a release in quite a while but that is because
  selector is simple and stable, not because it is unattended.
  (More below).

Is the library good code?

  I've been using selector in other codebases for several years
  and it does what it says on the tin with alacrity.

Is the library python 3 compatible?

  Yes.

Is the library license compatible?

  Selector is dual licensed GPL and MIT. The author states it
  may be used with either license. We choose MIT.

Is the library already packaged in the distros we target?

  No, but due to its limited surface area and minimal dependencies
  it should be straightforward.

Is the function of this library already covered by other libraries
in global-requirements.txt?

  As described above the Routes package provides similar functionality
  but in such a different style that it would be incorrect to call
  them equivalent.

Is the library required for OpenStack project or related dev or
infrastructure setup?

  Yes, as above, the plan is to use this with the new placement API
  being built in nova.

Code is at: https://github.com/lukearno/selector/
On PyPi at: https://pypi.python.org/pypi/selector

Change-Id: If1eb987fbe270e67a81005baa7f6f88c6586feef