Add selector to requirements
selector is a simple WSGI routing and delegation library that is
going to be used in the new placement API currently being built
in nova. selector is more lightweight, declarative, and
inspectable than the Routes library in common use throughout
OpenStack. There's an explicit goal in the placement API to
"use different stuff". selector is just one part of that.

It is dual licensed GPL and MIT
Code is at: https://github.com/lukearno/selector/
On PyPi at: https://pypi.python.org/pypi/selector
It works for both python 2 and 3.

There hasn't been a release in quite a while but that is because
selector is simple and stable, not because it is unattended.

Change-Id: If1eb987fbe270e67a81005baa7f6f88c6586feef